+++ src/latwuc.h
@@ -60,19 +60,20 @@ BOOL APIENTRY DllMain( HMODULE hModule,
 	return TRUE;
 }
 
-extern __declspec(dllexport)int tiffWrite(int width, int height, int pages,
+extern __declspec(dllexport)int tiffWrite(uint_least16_t width,
+                              uint_least16_t height, uint_least16_t pages,
                               char* artist, char* copyright, char* make,
                               char* model, char* software, char* image_desc,
                               char* output_path, uint16_t* buffer);
 #endif
 
 
-int writeSubFile(TIFF *image, int width, int height,
+int writeSubFile(TIFF *image, uint_least16_t width, uint_least16_t height,
                  char* artist, char* copyright, char* make, char* model,
                  char* software, char* image_desc,
                  uint16_t* buffer);
 
-int tiffWrite(int width, int height, int pages,
+int tiffWrite(uint_least16_t width, uint_least16_t height, uint_least16_t pages,
               char* artist, char* copyright, char* make, char* model,
               char* software, char* image_desc,
               char* output_path, uint16_t* buffer);
